+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/ABCInputStream.java
@@ -34,6 +34,7 @@ import com.jpexs.decompiler.flash.abc.types.traits.Traits;
 import com.jpexs.decompiler.flash.dumpview.DumpInfo;
 import com.jpexs.decompiler.flash.dumpview.DumpInfoSpecial;
 import com.jpexs.decompiler.flash.dumpview.DumpInfoSpecialType;
+import com.jpexs.helpers.Helper;
 import com.jpexs.helpers.MemoryInputStream;
 import com.jpexs.helpers.utf8.Utf8Helper;
 import java.io.ByteArrayOutputStream;
@@ -523,8 +524,8 @@ public class ABCInputStream implements AutoCloseable {
             stringDataBuffer = new byte[newLength];
         }
 
-        safeRead(length, stringDataBuffer);
-        String r = new String(stringDataBuffer, 0, length, Utf8Helper.charset);
+        safeRead(length, stringDataBuffer);                
+        String r = Utf8Helper.decode(stringDataBuffer, 0, length);
         endDumpLevel(r);
         return r;
     }